Create chia._tests.util.split_managers
