+++ chia/_tests/plot_sync/util.py
@@ -5,6 +5,7 @@ import time
 from dataclasses import dataclass
 from typing import AsyncIterator, Optional
 
+from chia._tests.util.split_managers import SplitAsyncManager, split_async_manager
 from chia._tests.util.time_out_assert import time_out_assert
 from chia.harvester.harvester import Harvester
 from chia.plot_sync.sender import Sender
@@ -14,7 +15,6 @@ from chia.types.aliases import FarmerService, HarvesterService
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo, UnresolvedPeerInfo
 from chia.util.ints import uint16, uint64
-from chia.util.misc import SplitAsyncManager, split_async_manager
 
 
 @dataclass
